키보드로 실수 입력 값 받아 출력하기

+++ learning/열혈강의C++/HelloWorld/SimpleAdder.cpp
@@ -2,15 +2,15 @@
 
 int main(void)
 {
-	int val1;
+	double val1;
 	std::cout << "첫 번째 숫자 입력: ";
-	std::cin >> val1; // 정수 입력 받아 val1에 저장
+	std::cin >> val1; // 실수 입력 받아 val1에 저장
 
-	int val2;
+	double val2;
 	std::cout << "두 번째 숫자 입력: ";
-	std::cin >> val2; // 정수 입력 받아 val2에 저장
+	std::cin >> val2; // 실수 입력 받아 val2에 저장
 
-	int result = val1 + val2;
+	double result = val1 + val2;
 	std::cout << "덧셈 결과: " << result << std::endl;
 
 	return 0;
